improve system architecture and input handling
- Add InputState struct for centralized input management
- Refactor DeathSystem to use proper entity cleanup with m_EntitiesToDelete vector
- Update collision, damage, and death systems with improved structure
- Add proper header includes and organize system dependencies

+++ game/src/systems/WaveSpawnerSystem.h
@@ -8,6 +8,6 @@ public:
     void update(World& world, float deltaTime) override;
 private:
     float _spawnTimer = 1.f;
-    float enemyNumber;
+    int enemyNumber;
     float _screenW, _screenH;
 };
